new: OIDC-backend parameters are now autodiscovered
The demo (both QT and testdefaultlicensing.cpp) now uses auto-discovery to populate the
OIDC-backend configuration. The user only needs to enter the base url and the autodiscovery
reads the OIDC-backend configuration automatically. Client configuration must still be
given.

The autodiscovery helps with configuration issues, in particular with the verification key.

Manual parameter-passing is still possible. E.g. the QT-form demoparametersdialog is still
around, but is not used.